Add files via upload
here's the sound effects I have so far, I'll add some for every object we need. The bag sound is three parts. The first is if you combine the bag and the other object sound effect, the second is just the bag sound and the last is a lower pitched version.